+++ config/config/src/Config.ts
@@ -160,6 +160,7 @@ export interface Config {
   dedupePeerDependents?: boolean
   patchesDir?: string
   ignoreWorkspaceCycles?: boolean
+  disallowWorkspaceCycles?: boolean
   packGzipLevel?: number
 
   registries: Registries
@@ -168,6 +169,7 @@ export interface Config {
 
   testPattern?: string[]
   changedFilesIgnorePattern?: string[]
+  rootProjectManifestDir?: string
   rootProjectManifest?: ProjectManifest
   userConfig: Record<string, string>
 }