+++ config/config/src/index.ts
@@ -25,6 +25,7 @@ import {
 } from './Config'
 import { getWorkspaceConcurrency } from './concurrency'
 
+export { getOptionsFromRootManifest, type OptionsFromRootManifest } from './getOptionsFromRootManifest'
 export * from './readLocalConfig'
 
 export type { Config, UniversalOptions }
@@ -45,6 +46,7 @@ export const types = Object.assign({
   'dedupe-direct-deps': Boolean,
   dev: [null, true],
   dir: String,
+  'disallow-workspace-cycles': Boolean,
   'enable-modules-dir': Boolean,
   'enable-pre-post-scripts': Boolean,
   'exclude-links-from-lockfile': Boolean,
@@ -193,6 +195,7 @@ export async function getConfig (
     'deploy-all-files': false,
     'dedupe-peer-dependents': true,
     'dedupe-direct-deps': false,
+    'disallow-workspace-cycles': false,
     'enable-modules-dir': true,
     'exclude-links-from-lockfile': false,
     'extend-node-path': true,
@@ -546,7 +549,8 @@ export async function getConfig (
   if (!pnpmConfig.ignorePnpmfile) {
     pnpmConfig.hooks = requireHooks(pnpmConfig.lockfileDir ?? pnpmConfig.dir, pnpmConfig)
   }
-  pnpmConfig.rootProjectManifest = await safeReadProjectManifestOnly(pnpmConfig.lockfileDir ?? pnpmConfig.workspaceDir ?? pnpmConfig.dir) ?? undefined
+  pnpmConfig.rootProjectManifestDir = pnpmConfig.lockfileDir ?? pnpmConfig.workspaceDir ?? pnpmConfig.dir
+  pnpmConfig.rootProjectManifest = await safeReadProjectManifestOnly(pnpmConfig.rootProjectManifestDir) ?? undefined
   if (pnpmConfig.rootProjectManifest?.workspaces?.length && !pnpmConfig.workspaceDir) {
     warnings.push('The "workspaces" field in package.json is not supported by pnpm. Create a "pnpm-workspace.yaml" file instead.')
   }