+++ patching/plugin-commands-patching/src/patchCommit.ts
@@ -42,7 +42,7 @@ export function help () {
   })
 }
 
-export async function handler (opts: install.InstallCommandOptions & Pick<Config, 'patchesDir' | 'rootProjectManifest'>, params: string[]) {
+export async function handler (opts: install.InstallCommandOptions & Pick<Config, 'patchesDir' | 'rootProjectManifest' | 'rootProjectManifestDir'>, params: string[]) {
   const userDir = params[0]
   const lockfileDir = opts.lockfileDir ?? opts.dir ?? process.cwd()
   const patchesDirName = normalizePath(path.normalize(opts.patchesDir ?? 'patches'))
@@ -86,6 +86,7 @@ export async function handler (opts: install.InstallCommandOptions & Pick<Config
 
   return install.handler({
     ...opts,
+    rootProjectManifest,
     rawLocalConfig: {
       ...opts.rawLocalConfig,
       'frozen-lockfile': false,