+++ pkg-manager/core/src/install/extendInstallOptions.ts
@@ -63,6 +63,7 @@ export interface StrictInstallOptions {
   engineStrict: boolean
   neverBuiltDependencies?: string[]
   onlyBuiltDependencies?: string[]
+  onlyBuiltDependenciesFile?: string
   nodeExecPath?: string
   nodeLinker: 'isolated' | 'hoisted' | 'pnp'
   nodeVersion: string
@@ -105,6 +106,7 @@ export interface StrictInstallOptions {
   resolutionMode: 'highest' | 'time-based' | 'lowest-direct'
   resolvePeersFromWorkspaceRoot: boolean
   ignoreWorkspaceCycles: boolean
+  disallowWorkspaceCycles: boolean
 
   publicHoistPattern: string[] | undefined
   hoistPattern: string[] | undefined
@@ -223,6 +225,7 @@ const defaults = (opts: InstallOptions) => {
     resolvePeersFromWorkspaceRoot: true,
     extendNodePath: true,
     ignoreWorkspaceCycles: false,
+    disallowWorkspaceCycles: false,
     excludeLinksFromLockfile: false,
   } as StrictInstallOptions
 }