+++ pkg-manager/headless/src/index.ts
@@ -1,6 +1,7 @@
 import { promises as fs } from 'fs'
 import path from 'path'
 import { buildModules } from '@pnpm/build-modules'
+import { createAllowBuildFunction } from '@pnpm/builder.policy'
 import { calcDepState, type DepsStateCache } from '@pnpm/calc-dep-state'
 import {
   LAYOUT_VERSION,
@@ -93,6 +94,9 @@ export interface Project {
 }
 
 export interface HeadlessOptions {
+  neverBuiltDependencies?: string[]
+  onlyBuiltDependencies?: string[]
+  onlyBuiltDependenciesFile?: string
   autoInstallPeers?: boolean
   childConcurrency?: number
   currentLockfile?: Lockfile
@@ -488,6 +492,7 @@ export async function headlessInstall (opts: HeadlessOptions): Promise<Installat
       }
     }
     await buildModules(graph, Array.from(directNodes), {
+      allowBuild: createAllowBuildFunction(opts),
       childConcurrency: opts.childConcurrency,
       extraBinPaths,
       extraEnv,