+++ cli/cli-utils/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @pnpm/cli-utils
 
+## 2.0.24
+
+### Patch Changes
+
+- Updated dependencies [ac5abd3ff]
+- Updated dependencies [b60bb6cbe]
+  - @pnpm/config@20.0.0
+  - @pnpm/default-reporter@12.4.3
+
+## 2.0.23
+
+### Patch Changes
+
+- Updated dependencies [b1dd0ee58]
+  - @pnpm/config@19.2.1
+  - @pnpm/default-reporter@12.4.2
+
+## 2.0.22
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+- Updated dependencies [d774a3196]
+- Updated dependencies [832e28826]
+  - @pnpm/config@19.2.0
+  - @pnpm/types@9.3.0
+  - @pnpm/default-reporter@12.4.1
+  - @pnpm/cli-meta@5.0.3
+  - @pnpm/package-is-installable@8.0.5
+  - @pnpm/manifest-utils@5.0.4
+  - @pnpm/read-project-manifest@5.0.7
+
 ## 2.0.21
 
 ### Patch Changes