+++ hooks/types/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/hooks.types
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+  - @pnpm/types@9.3.0
+  - @pnpm/lockfile-types@5.1.2
+
 ## 1.0.2
 
 ### Patch Changes