+++ packages/core-loggers/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pnpm/core-loggers
 
+## 9.0.3
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+  - @pnpm/types@9.3.0
+
 ## 9.0.2
 
 ### Patch Changes