+++ packages/plugin-commands-init/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @pnpm/plugin-commands-init
 
+## 3.0.24
+
+### Patch Changes
+
+- Updated dependencies [ac5abd3ff]
+- Updated dependencies [b60bb6cbe]
+  - @pnpm/config@20.0.0
+  - @pnpm/cli-utils@2.0.24
+
+## 3.0.23
+
+### Patch Changes
+
+- Updated dependencies [b1dd0ee58]
+  - @pnpm/config@19.2.1
+  - @pnpm/cli-utils@2.0.23
+
+## 3.0.22
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+- Updated dependencies [832e28826]
+  - @pnpm/config@19.2.0
+  - @pnpm/cli-utils@2.0.22
+  - @pnpm/write-project-manifest@5.0.3
+
 ## 3.0.21
 
 ### Patch Changes