+++ reviewing/plugin-commands-licenses/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @pnpm/plugin-commands-licenses
 
+## 2.0.30
+
+### Patch Changes
+
+- Updated dependencies [ac5abd3ff]
+- Updated dependencies [b60bb6cbe]
+  - @pnpm/config@20.0.0
+  - @pnpm/license-scanner@2.0.22
+  - @pnpm/cli-utils@2.0.24
+
+## 2.0.29
+
+### Patch Changes
+
+- Updated dependencies [b1dd0ee58]
+  - @pnpm/config@19.2.1
+  - @pnpm/cli-utils@2.0.23
+
+## 2.0.28
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+- Updated dependencies [832e28826]
+  - @pnpm/config@19.2.0
+  - @pnpm/cli-utils@2.0.22
+  - @pnpm/lockfile-file@8.1.3
+  - @pnpm/license-scanner@2.0.21
+
 ## 2.0.27
 
 ### Patch Changes