+++ store/plugin-commands-store/CHANGELOG.md
@@ -1,5 +1,63 @@
 # @pnpm/plugin-commands-store
 
+## 8.0.29
+
+### Patch Changes
+
+- @pnpm/store-connection-manager@7.0.8
+
+## 8.0.28
+
+### Patch Changes
+
+- Updated dependencies [01bc58e2c]
+- Updated dependencies [ac5abd3ff]
+- Updated dependencies [b60bb6cbe]
+  - @pnpm/store.cafs@2.0.6
+  - @pnpm/config@20.0.0
+  - @pnpm/store-connection-manager@7.0.7
+  - @pnpm/cli-utils@2.0.24
+
+## 8.0.27
+
+### Patch Changes
+
+- @pnpm/store-connection-manager@7.0.6
+
+## 8.0.26
+
+### Patch Changes
+
+- @pnpm/store-connection-manager@7.0.5
+
+## 8.0.25
+
+### Patch Changes
+
+- Updated dependencies [b1dd0ee58]
+  - @pnpm/config@19.2.1
+  - @pnpm/cli-utils@2.0.23
+  - @pnpm/store-connection-manager@7.0.4
+
+## 8.0.24
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+- Updated dependencies [d774a3196]
+- Updated dependencies [832e28826]
+  - @pnpm/config@19.2.0
+  - @pnpm/types@9.3.0
+  - @pnpm/cli-utils@2.0.22
+  - @pnpm/store-connection-manager@7.0.3
+  - @pnpm/normalize-registries@5.0.3
+  - @pnpm/pick-registry-for-package@5.0.3
+  - @pnpm/lockfile-utils@8.0.6
+  - @pnpm/dependency-path@2.1.4
+  - @pnpm/get-context@10.0.4
+  - @pnpm/store.cafs@2.0.5
+  - @pnpm/store-controller-types@17.0.1
+
 ## 8.0.23
 
 ### Patch Changes