+++ store/store-controller-types/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pnpm/store-controller-types
 
+## 17.0.1
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+  - @pnpm/types@9.3.0
+  - @pnpm/fetcher-base@15.0.2
+  - @pnpm/resolver-base@10.0.3
+
 ## 17.0.0
 
 ### Major Changes