+++ workspace/filter-workspace-packages/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @pnpm/filter-workspace-packages
 
+## 7.1.4
+
+### Patch Changes
+
+- @pnpm/workspace.pkgs-graph@2.0.8
+- @pnpm/workspace.find-packages@1.0.14
+
+## 7.1.3
+
+### Patch Changes
+
+- @pnpm/workspace.find-packages@1.0.13
+
+## 7.1.2
+
+### Patch Changes
+
+- @pnpm/workspace.find-packages@1.0.12
+- @pnpm/workspace.pkgs-graph@2.0.7
+
 ## 7.1.1
 
 ### Patch Changes