+++ workspace/find-packages/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @pnpm/find-workspace-packages
 
+## 1.0.14
+
+### Patch Changes
+
+- @pnpm/cli-utils@2.0.24
+
+## 1.0.13
+
+### Patch Changes
+
+- @pnpm/cli-utils@2.0.23
+
+## 1.0.12
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+  - @pnpm/types@9.3.0
+  - @pnpm/cli-utils@2.0.22
+  - @pnpm/fs.find-packages@2.0.7
+
 ## 1.0.11
 
 ### Patch Changes